Remove dependecy from gdal , user rasterio for georeferencing

+++ predictor/app.py
@@ -12,7 +12,13 @@ from .vectorizer import vectorize
 
 
 def predict(
-    bbox, model_path, zoom_level, tms_url, tile_size=256, use_raster2polygon=False
+    bbox,
+    model_path,
+    zoom_level,
+    tms_url,
+    tile_size=256,
+    use_raster2polygon=False,
+    remove_metadata=True,
 ):
     """
     Parameters:
@@ -66,5 +72,6 @@ def predict(
     print(f"It took {round(time.time()-start)} sec to extract polygons")
     with open(geojson_path, "r") as f:
         prediction_geojson_data = json.load(f)
-    shutil.rmtree(base_path)
+    if remove_metadata:
+        shutil.rmtree(base_path)
     return prediction_geojson_data
